Minor fixes plus new configuration setting
Fixed failing tests caused by accuracy of last modified time being
returned to nearest millisecond in FileProps but HTTP header only
contains time to nearest second.

Added new configuration setting redirect_404_to_index which will return
index.html if the requested file is not found. This is needed for
AngularJS applications as the URLs are used by angular on the
client-side to make the application show the correct page and will not
exist on the server, in this situation the server needs to return the
index page.

Fixed error on Windows where paths were not being combined correctly so
file could not be found.